contuniacion User 3

+++ wms/src/main/java/big_three/wms/service/UserService.java
@@ -2,12 +2,15 @@ package big_three.wms.service;
 
 import big_three.wms.model.User;
 import big_three.wms.dto.UserCreateDTO;
+import big_three.wms.dto.UserResponseDTO;
 import big_three.wms.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
+
 
 @Service
 public class UserService {
@@ -20,7 +23,7 @@ public class UserService {
     public User create(UserCreateDTO dto) {
         //validaciones de Negocio
         if (userRepository.existsByCuil(dto.getCuil())) {
-            throw new IllegalArgumentException("Ya existe un usuario con ese CUIL");
+            throw new IllegalArgumentException("Ya existe un usuario con ese CUIL" + dto.getCuil());
         }
 
         User u = new User();
@@ -28,19 +31,39 @@ public class UserService {
         u.setApellido(dto.getApellido());
         u.setCuil(dto.getCuil());
         u.setContrasena(passwordEncoder.encode(dto.getContrasena()));
-        return userRepository.save(u);
+        User savedUser= userRepository.save(u);
+        return convertToResponseDTO(savedUser);
+    }
+    // Retorna una lista de DTOs mapeados
+    public List<UserResponseDTO> findAll() {
+        return userRepository.findAll()
+                .stream()
+                .map(this::convertToResponseDTO)
+                .collect(Collectors.toList());
     }
 
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-
-    public User findById(UUID id) {
-        return userRepository.findById(id)
+    // Retorna un DTO de respuesta o lanza error
+    public UserResponseDTO findById(UUID id) {
+        User u = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+        return convertToResponseDTO(u);
     }
 
     public void deleteById(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("Usuario no encontrado para eliminar");
+        }
         userRepository.deleteById(id);
+    }
+
+
+private UserResponseDTO convertToResponseDTO(User user) {
+    UserResponseDTO response = new UserResponseDTO();
+    response.setIdUsuario(user.getIdUsuario());
+    response.setNombre(user.getNombre());
+    response.setApellido(user.getApellido());
+    response.setCuil(user.getCuil());
+    response.setRol(user.getRol());
+    return response;
     }
 }
